% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Discord.R
\name{Add_Discord}
\alias{Add_Discord}
\title{Add_Discord}
\usage{
Add_Discord(dataset)
}
\arguments{
\item{dataset}{A csv file with data from each sites.
The first column is the Years.
The second column is the Months (from 1 to 12).
The third column is the data coming from each site.
See 'Dataset.csv' as example.}
}
\value{
Hosking and Wallis' discordancy measure
calculated using the additive approach as proposed in
'LOPES MARTINS et al. (2022)'
https://doi.org/10.1590/1678-4499.20220061
}
\description{
Add_Discord
}
\examples{
data(dataset)
Add_Discord(dataset)
}
