% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Heterogenety.R
\name{Add_Heterogenety}
\alias{Add_Heterogenety}
\title{Add_Heterogenety}
\usage{
Add_Heterogenety(dataset, rho, Ns = 100)
}
\arguments{
\item{dataset}{A csv file with data from each sites.
The first column is the Years.
The second column is the Months (from 1 to 12).
The third column is the data coming from each site.
Use data(dataset) as example.}

\item{rho}{A 1-column csv file or matrix
with 12 monthly values for
coefficients of spatial correlation.
The first coefficient is for January.
The last coefficient is for December.}

\item{Ns}{Number of simulated groups of series.
Default is 100, but at least 500 is recommended.}
}
\value{
'Hosking' and 'Wallis'' heterogeneity measure
calculated using the additive approach proposed in
'LOPES MARTINS et al. (2022)'
https://doi.org/10.1590/1678-4499.20220061
}
\description{
Add_Heterogenety
}
\examples{
data(dataset)
rho=as.matrix(rep(0.5,12))
Add_Heterogenety(dataset,rho = rho,Ns = 100)
}
