% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegPar.R
\name{Add_RegPar}
\alias{Add_RegPar}
\title{Add_RegPar}
\usage{
Add_RegPar(dataset)
}
\arguments{
\item{dataset}{A csv file with data from each sites.
The first column is the Years.
The second column is the Months (from 1 to 12).
The third column is the data coming from each site.
See 'Dataset.csv' as example.}
}
\value{
Regional parameter for several distributions
GEV: Generalized extreme value distribution.
GLO: Generalized logistic distribution.
PE3: Pearson type III distribution.
GNO: Generalized normal distribution.
GPA: Generalized pareto distribution.
}
\description{
Add_RegPar
}
\examples{
data(dataset)
Add_RegPar(dataset)
}
