% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegProb.R
\name{Add_RegProb}
\alias{Add_RegProb}
\title{Add_RegProb}
\usage{
Add_RegProb(dataset, distr, bestPar)
}
\arguments{
\item{dataset}{A csv file with data from each sites.
The first column is the Years.
The second column is the Months (from 1 to 12).
The third column is the data coming from each site.
Use data(dataset) as example.}

\item{distr}{A string indicating the regional distribution
\dQuote{GEV} \dQuote{GLO} \dQuote{PE3}
\dQuote{GNO} \dQuote{GPA}.}

\item{bestPar}{Parameters of the regional distribution (selected by 'Add_Goodness').
Obtained from 'Add_RegPar'. See example.}
}
\value{
Cumulative probabilities for the sites forming
the homogeneous region or group.
}
\description{
Add_RegProb
}
\examples{
data(dataset)
data(parGEV)
Add_RegProb(dataset,distr="GEV",bestPar=parGEV)
}
