% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_RegSPEI.R
\name{Add_RegSPEI}
\alias{Add_RegSPEI}
\title{Add_RegSPEI}
\usage{
Add_RegSPEI(RegProb)
}
\arguments{
\item{RegProb}{The cumulative probabilities for the sites forming
the homogeneous region or group.
Calculated using the 'Add_RegProb' function.}
}
\value{
The Regional SPEI
}
\description{
Add_RegSPEI
}
\examples{
data(RegProb)
RegSPEI=Add_RegSPEI(RegProb)
}
