% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Difference between precipitation and 'evapotranspiration'}
\format{
A \code{data.frame} with 15 variables and 344 rows.

\describe{
\item{Year}{Year}
\item{Month}{Month}
\item{site1}{precipitation minus 'evapotranspiration' at site 1}
\item{site2}{precipitation minus 'evapotranspiration' at site 2}
\item{site3}{precipitation minus 'evapotranspiration' at site 3}
\item{site4}{precipitation minus 'evapotranspiration' at site 4}
\item{site5}{precipitation minus 'evapotranspiration' at site 5}
\item{site6}{precipitation minus 'evapotranspiration' at site 6}
\item{site7}{precipitation minus 'evapotranspiration' at site 7}
\item{site8}{precipitation minus 'evapotranspiration' at site 8}
\item{site9}{precipitation minus 'evapotranspiration' at site 9}
\item{site10}{precipitation minus 'evapotranspiration' at site 10}
\item{site11}{precipitation minus 'evapotranspiration' at site 11}
\item{site12}{precipitation minus 'evapotranspiration' at site 12}
\item{site13}{precipitation minus 'evapotranspiration' at site 13}
}
}
\source{
Generated by the statistical simulation using the
generalized extreme value distribution.
}
\usage{
dataset
}
\description{
Contains monthly values for the difference
between precipitation and 'evapotranspiration' (1993 to 2021).
}
\examples{
data(dataset)
}
\keyword{datasets}
