% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parGEV.R
\docType{data}
\name{parGEV}
\alias{parGEV}
\title{Regional parameters of the generalized extreme value distribution}
\format{
A \code{data.frame} with 4 variables and 12 rows.

\describe{
\item{Month}{Month}
\item{loc}{location parameter}
\item{scl}{scale parameter}
\item{shp}{shape parameter}
}
}
\source{
Generated by the statistical simulation using the
generalized extreme value distribution.
}
\usage{
parGEV
}
\description{
Contains parameters of the generalized extreme value distribution
fitted through the additive regional frequency analysis.
}
\examples{
data(parGEV)
}
\keyword{datasets}
