% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegProb.R
\docType{data}
\name{RegProb}
\alias{RegProb}
\title{Cumulative probabilities of the precipitation minus 'evapotranspiration'}
\format{
A \code{data.frame} with 15 variables and 344 rows.

\describe{
\item{Year}{Year}
\item{Month}{Month}
\item{site1}{probabilities of precipitation minus 'evapotranspiration' at site 1}
\item{site2}{probabilities of precipitation minus 'evapotranspiration' at site 2}
\item{site3}{probabilities of precipitation minus 'evapotranspiration' at site 3}
\item{site4}{probabilities of precipitation minus 'evapotranspiration' at site 4}
\item{site5}{probabilities of precipitation minus 'evapotranspiration' at site 5}
\item{site6}{probabilities of probabilities of precipitation minus 'evapotranspiration' at site 6}
\item{site7}{probabilities of precipitation minus 'evapotranspiration' at site 7}
\item{site8}{probabilities of precipitation minus 'evapotranspiration' at site 8}
\item{site9}{probabilities of precipitation minus 'evapotranspiration' at site 9}
\item{site10}{probabilities of precipitation minus 'evapotranspiration' at site 10}
\item{site11}{probabilities of precipitation minus 'evapotranspiration' at site 11}
\item{site12}{probabilities of precipitation minus 'evapotranspiration' at site 12}
\item{site13}{probabilities of precipitation minus 'evapotranspiration' at site 13}
}
}
\source{
Generated by the statistical simulation using the
generalized extreme value distribution with regional parameters.
}
\usage{
RegProb
}
\description{
Contains monthly values for the cumulative probabilities
of precipitation minus 'evapotranspiration' totals (1993 to 2021).
}
\examples{
data(RegProb)
}
\keyword{datasets}
