% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Goodness.R
\name{Add_GoodnessFit}
\alias{Add_GoodnessFit}
\title{Add_GoodnessFit}
\usage{
Add_GoodnessFit(dataset, rho = 0, Ns = 100)
}
\arguments{
\item{dataset}{A csv file with data from each sites.
The first column is the Years.
The second column is the Months (from 1 to 12).
The third column is the data coming from each site.
Use data(dataset) as example.}

\item{rho}{Average spatial correlation (between 0 and 1).
defauld is 0.}

\item{Ns}{Number of simulated groups of series.
Default is 100, but at least 500 is recommended.}
}
\value{
Hosking and Wallis (1991) Goodness of fit measure and
its bivatriate extension proposed in Kjeldsen and Prosdocimi (2015).
}
\description{
Add_GoodnessFit
}
\examples{
data(dataset)
Add_GoodnessFit(dataset,rho=0.5,Ns=100)
}
