% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_Heterogenety.R
\name{Add_Heterogenety}
\alias{Add_Heterogenety}
\title{Add_Heterogenety}
\usage{
Add_Heterogenety(dataset, rho = 0, Ns = 100)
}
\arguments{
\item{dataset}{A csv file with data from each sites.
The first column is the Years.
The second column is the Months (from 1 to 12).
The third column is the data coming from each site.
See 'Dataset.csv' as example.}

\item{rho}{Average spatial correlation (between 0 and 1).
defauld is 0.}

\item{Ns}{Number of simulated groups of series.
Default is 100, but at least 500 is recommended.}
}
\value{
'Hosking' and 'Wallis' Goodness of fit measure
'Kjeldsen' and 'Prosdocimi's' bivariate extension
calculated using the additive approach proposed in
'LOPES MARTINS et al.' (2022)
https://doi.org/10.1590/1678-4499.20220061

'Hosking' and 'Wallis'' heterogeneity measure
calculated using the additive approach proposed in
'LOPES MARTINS et al. (2022)'
https://doi.org/10.1590/1678-4499.20220061
}
\description{
Add_Heterogenety
}
\examples{
data(dataset)
Add_Heterogenety(dataset,rho = 0.5,Ns = 100)
}
